backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
imagekitio==3.2.0
bleach==6.1.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
python-slugify==8.0.1
